+++ backend-master/lib/models.rb
@@ -1,4 +1,3 @@
 # frozen_string_literal: true
 
-require_relative 'models/user'
 require_relative 'models/order'