+++ backend-master/lib/market/btc_market.rb
@@ -4,7 +4,7 @@ module PaymiumMarket
   module Market
     class BTCMarket < PaymiumMarket::Market::Base
       # sets the database connector
-      def initialize(db_connector = PaymiumMarket::Database::MysqlConnector.new("#{BASE}/#{QUOTE}"), fee = 0.25, fee_user_id = 1)
+      def initialize(db_connector = PaymiumMarket::Database::MysqlConnector.new('BTC/EUR'), fee = 0.25, fee_user_id = 1)
         @db = db_connector
         @base = 'btc'
         @quote = 'eur'